requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
python-dotenv==1.0.1
gunicorn==25.1.0

pandas==2.2.2
numpy==1.26.4
scikit-learn==1.8.0

openai==1.14.3

sentence-transformers==2.6.1
transformers==4.57.6
torch==2.2.2

faiss-cpu==1.8.0

requests==2.32.5
tqdm==4.66.2